// openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/barcode-scanners/scandit-scanner-cordova/types/barcode.ts
export enum Symbology {
    EAN13UPCA = 'ean13Upca',
    UPCE = 'upce',
    EAN8 = 'ean8',
    Code39 = 'code39',
    Code93 = 'code93',
    Code128 = 'code128',
    Code11 = 'code11',
    Code25 = 'code25',
    Codabar = 'codabar',
    InterleavedTwoOfFive = 'interleavedTwoOfFive',
    ITF = 'itf',
    MSIPlessey = 'msiPlessey',
    QR = 'qr',
    DataMatrix = 'dataMatrix',
    Aztec = 'aztec',
    MaxiCode = 'maxicode',
    DotCode = 'dotcode',
    KIX = 'kix',
    RM4SCC = 'rm4scc',
    GS1Databar = 'databar',
    GS1DatabarExpanded = 'databarExpanded',
    GS1DatabarLimited = 'databarLimited',
    PDF417 = 'pdf417',
    MicroPDF417 = 'microPdf417',
    MicroQR = 'microQr',
    Code32 = 'code32',
    Lapa4SC = 'lapa4sc',
}